worker/requirements.txt
SQLAlchemy~=2.0.44
psycopg2-binary~=2.9.11
nats-py~=2.11.0
pydantic~=2.12.3
pydantic-settings~=2.11.0
